Fadelabel group test
This adds a group of fadelabels with different weather fields as a test
for a rotating future forecast overlay piece. The controls do change in
sync. The issue with not redrawing the controls does appear to be linked
to not forcing a refresh. If "Ken Burns" pan & zoom effects are turned
on, the labels look normal. If they are not, the fadelabels only update
when the picture changes and the rest of the screen is refreshed. Next
step will be to add a multi image control for outlook icons and attempt
to sync it with the fadelabels.