Overlay now works for both video and pictures
Our overlay now works for mixed media video and picture playlists in the
way we use it (to the extent I can manually test it). The bulk of the
changes come from rebasing the skin on the latest version of confluence
which includes a number of new Gotham changes. In some cases, I have
made those changes in our portions of the code. In other cases,
specifically where posx is being used with right align, I've left posx
because changing to left causes layout problems I have not had time to
fix. The video-specific overlay changes are in VideoFullScreen.xml.
There's still more to do, but this should see us through the demo,
barring the appearance of any bugs/issues.